# all: reduce string allocations