Before:
```
BenchmarkLargeWEBWrite-16    	    1886	    612136 ns/op	   89070 B/op	    2659 allocs/op
PASS
ok  	github.com/moov-io/ach	3.844s
```

After: 
```
BenchmarkLargeWEBWrite-16    	    2000	    566199 ns/op	   54180 B/op	    2041 allocs/op
PASS
ok  	github.com/moov-io/ach	3.210s
```